Ref [https://trac.osgeo.org/fusion/ticket/199] - Apply selection speed improvements to fusion2-mg21 branch
Author:           Ronnie Louie
Date:             Thursday, December 18, 2008
Title:            1161170 Improve selection performance in Fusion
Reviewed by:      Chris Claydon

========
Abstract
========
Selection performance in Fusion is significantly slower than the performance on 
the Ajax viewer with basic weblayouts.  It generally can take two or three times 
as long (if not more) to select features in Flexible WebLayouts than in Basic 
layouts.  The slowdown is attributed to the additional server requests that are 
made when the Query.php script is executed for Fusion.  To improve the response 
time, feedback is provided to the user by highlighting the selection on the 
screen first, via the quicker QueryMapFeatures request, and then feature 
attributes for the selection are retrieved and displayed.  The end result is an 
improved user experience when selecting features in flexible weblayout.  Running 
in a debug environment (server and webtier) selection of roughly 1500 features 
would take around 8 seconds before the selection is displayed.  It now takes 
about 3 seconds to display.

TESTING PERFORMED:
- Tested point-selection and drag-selection for a single selectable layer.
- Tested point-selection and drag-selection for multiple selectable layers.
- Tested extended-selection via point and drag select.
- Tested toggle on/off of features when the features are selected again.
- Tested radius/polygon select for single and multiple selectable layers.
- Tested radius/polygon select for extended selection.
- Tested radius/polygon select for toggling selected features.
- Tested selection for all layers.
- Tested selection for only the active layer.
- Tested zoom selection.
- All tests executed in FireFox and Internet Explorer.

================
Module by Module
================
//MgDev/OS/Oem/fusion/layers/MapGuide/MapGuide.js
  - modified selection to draw the overlay first before retrieving the 
attributes to display

//MgDev/OS/Oem/fusion/layers/MapGuide/php/HighlightSelection.php
  - new file added to save the selection to the repository and return a json 
data structure 

//MgDev/OS/Oem/fusion/layers/MapGuide/php/SetSelection.php
  - removed saving of the selection set to the repository since this is already 
done in HightlightSelection.php
  - optimize the code so that the retrieval of spatial contexts and creation of 
a coordinate system is moved into in if block to be executed only when computed 
properties are needed.  The existing code was not doing anything for computed 
properties so this block of coded did not need to be executed.
  - replace the code for determining the feature count for each layer with a 
call to MgSelection::GetSelectedFeaturesCount()


git-svn-id: http://svn.osgeo.org/fusion/branches/fusion2-mg21@1727 1f2f6c64-2f36-0410-af3b-c496fd8ebf2b