re [https://trac.osgeo.org/fusion/ticket/199] - Fix setSelection API to work with new selection handling.
MapGuide.js - cleaned up the code a little, and re-structured so that setSelection updates the map image, and optionally zooms to the extents of the selected features.

Renamed HighlightSelection.php to SaveSelection.php, since all it does is update the selection stored in the map resource in the MG session, and retrieve some basic info about the selection.
Renamed SetSelection.php to GetSelectionProperties.php, since it no longer sets the selection. It just retrieves properties and basic info for the selected features.

Also updated SaveSelection.php optionally to retrieve the selection extents. This is required is setSelection is called with zoomTo set to true.

git-svn-id: http://svn.osgeo.org/fusion/branches/fusion2-mg21@1729 1f2f6c64-2f36-0410-af3b-c496fd8ebf2b