Fixed ticket http://trac.osgeo.org/fusion/ticket/337 Fusion: Select Within doesn't work.
Comments from Buddy who fixed a similar defect in MapGuide:
In this patch, I check the type for each feature. And create a multipolygon if there are only polygon. If there are unsupported types, continue and go to the next one. For mixed supported types, create a multi-geometry. 


BTW, the MG Server, the render service doesn't support QueryFeatures?? for the multi-geometry. So If we create a multi-polygoemetry, QueryFeatures??. will return null, and everything in the map will be cleared. 


We need address another defect for this issue. 



git-svn-id: http://svn.osgeo.org/fusion/trunk@2026 1f2f6c64-2f36-0410-af3b-c496fd8ebf2b