Fix ticket [https://trac.osgeo.org/fusion/ticket/511]: Error message when layer name starting with number in MapGuide
When Fusion try to store selection set in session, it uses layer name as property name, but PHP doesn't allow property name beginning with number. That's why. The fix is adding a prefix "Layer" before the layer name when using it as property name. When fetching data from session, we remove the prefix so that there is no change to the result.


git-svn-id: http://svn.osgeo.org/fusion/trunk@2512 1f2f6c64-2f36-0410-af3b-c496fd8ebf2b