Submit on behalf of Andy Zhang. Integrate 2650 to trunk.
Because the widget GoogleStreetView changed the error handler of undefined CS. For the CSs which are not defined in fusion, now it throws an exception. While it just skips the error in 2013. And fusion only defines a few CSs, this will result in many users get exception dialog when using fusion web layout.
To fix it, we need to change GoogleStreetView.js to only throw exception when GoogleStreetView is activated.


git-svn-id: http://svn.osgeo.org/fusion/trunk@2653 1f2f6c64-2f36-0410-af3b-c496fd8ebf2b