Submit on behalf of Andy Zhang. Integrate [https://trac.osgeo.org/fusion/ticket/2640] to trunk.
Set default value of DisplayProjection of widget CursorPosition to null. 
The default value of DisplayProjection widget CursorPosition is EPSG:4326. If we don't change it manually, the cursor position is always displayed in EPSG:4326 format, which is not correct. Now we don't set the default value. The cursor position widget will get projection from map. Then the x, y value of cursor position is consistent with map CS.

git-svn-id: http://svn.osgeo.org/fusion/trunk@2662 1f2f6c64-2f36-0410-af3b-c496fd8ebf2b