[https://trac.osgeo.org/fusion/ticket/565]: Add 2 new extension parameters to the Redline widget:
 - DataStoreFormat: The redline data store format to use (SDF/SHP/SQLite)
 - RedlineGeometryFormat: The types of geometry the user can record with this widget (1=Point, 2=Line, 4=Geometry and any bitmask combination thereof)

If both values are specified (and only a singular geometry type if the format is SHP), then the "Create New Redline" section of the UI is replaced with just a button of the same name. This insulates the user from needing to know what format to use and what geometry types to tick as this can be imposed by the author of the Flexible Layout.

git-svn-id: http://svn.osgeo.org/fusion/trunk@2712 1f2f6c64-2f36-0410-af3b-c496fd8ebf2b