[https://trac.osgeo.org/fusion/ticket/598]: Use closure compiler for compressing our javascript. Besides producing smaller files (50kb less) on default settings, closure gives us sourcemap support which will make "debugging" fusionSF-compressed.js a much more pleasant affair. We're still keeping YUICompressor around for any CSS compression tasks.
Note: We're not using closure with advanced optimizations. Although advanced optimizations will make the resulting file 500kb smaller, it may change the actual behaviour of our code. To move to advanced optimizations requires a thorough review of the code base to determine if it is compatible with the closure compiler in this optimization mode.

git-svn-id: http://svn.osgeo.org/fusion/trunk@2796 1f2f6c64-2f36-0410-af3b-c496fd8ebf2b