Merged revision(s) 2969-2972 from sandbox/adsk/3.2o: Ticket 2763
The javascript method - createSessionId in fusion.js will send the HTTP  request with  the "GET" method. The user name and the password are  visible. So the "GET" method is changed to the "POST" method.
........
[https://trac.osgeo.org/fusion/ticket/656] Error when select redline object.

Not quite sure why the error happens, just add some checks to avoid error.
........
Use OpenLayers.Layer.Bing instead of deprecated OpenLayers.Layer.VirtualEarth.

Microsoft will not support Bing map v6/v7 in the middle of 2017. As a result, we have to use OpenLayers.Layer.Bing for Bing map. The new class uses Bing map tile service, which an application key is required. We will store AIMS application key in application definition xml, and read the key when opening the xml.

This submission also fixed 2 small issues of flexible web layout.
1. Update the INCH_PER_METER to 39.370078, which is more precise.
2. Use 'post' in MgQueryMapFeatures2 because the URL may be too long.

........
Handle the case that the map hasn't a CS assigned in widget SelectWithin.
........


git-svn-id: http://svn.osgeo.org/fusion/trunk@2993 1f2f6c64-2f36-0410-af3b-c496fd8ebf2b